+++ packages/components/src/Banner/notificationTypes.css.d.ts
@@ -1,8 +1,0 @@
-declare const styles: {
-  readonly "notice": string;
-  readonly "success": string;
-  readonly "warning": string;
-  readonly "error": string;
-};
-export = styles;
-