+++ packages/components/src/Box/styles/BoxDimension.css.d.ts
@@ -1,0 +1,10 @@
+declare const styles: {
+  readonly "height-auto": string;
+  readonly "width-auto": string;
+  readonly "height-shrink": string;
+  readonly "width-shrink": string;
+  readonly "height-grow": string;
+  readonly "width-grow": string;
+};
+export = styles;
+