+++ packages/components/src/FormField/hooks/useIsSafari.ts
@@ -1,0 +1,5 @@
+export function useIsSafari(): boolean {
+  return globalThis?.navigator
+    ? /^((?!chrome|android).)*safari/i.test(navigator.userAgent)
+    : false;
+}