+++ packages/hooks/src/index.ts
@@ -13,3 +13,4 @@ export * from "./useRefocusOnActivator";
 export * from "./useResizeObserver";
 export * from "./useSafeLayoutEffect";
 export * from "./useShowClear";
+export * from "./useWindowDimensions";